Renaming components packages to lowercase and move the BarchartV package to a generic "barchart" package
See: https://github.com/npm/npm, issue 7351

Close #31
