--- packages/components/BarchartV/CHANGELOG.md
@@ -1,3 +1,0 @@
-## `@svizzle/BarchartV` v0.1.0 (next)
-
-- added the component
