+++ packages/components/barchart/CHANGELOG.md
@@ -1,0 +1,3 @@
+## `@svizzle/barchart` v0.1.0
+
+- added the component
