+++ packages/components/barchart/LICENSE
