+++ packages/components/barchart/src/BarchartV.svelte
