+++ packages/tools/atlas/CHANGELOG.md
@@ -1,4 +1,4 @@
 ## `@svizzle/atlas` v0.1.0
 
-- moved world boundaries scripts from /Choropleth
+- moved world boundaries scripts from /choropleth
 - add NUTS scripts
